Move inputs to session
