+++ app/src/main/java/dev/upaya/shf/ui/session/noting/NotingScreen.kt
@@ -15,7 +15,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import dev.upaya.shf.exercises.labels.Label
 import dev.upaya.shf.ui.KeepScreenOn
 import dev.upaya.shf.ui.SetStatusBarColor
-import dev.upaya.shf.ui.input.InputViewModel
+import dev.upaya.shf.ui.session.input.InputViewModel
 import dev.upaya.shf.ui.session.LabelText
 import dev.upaya.shf.ui.session.SessionViewModel
 import dev.upaya.shf.ui.simulatePress
