+++ app/src/main/java/dev/upaya/shf/ui/session/noting/stats/composables/LabelStatsCard.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import dev.upaya.shf.exercises.labels.labelFeel
 import dev.upaya.shf.exercises.labels.labelHear
 import dev.upaya.shf.exercises.labels.labelSee
-import dev.upaya.shf.ui.input.LabelFreqs
+import dev.upaya.shf.ui.session.input.LabelFreqs
 import dev.upaya.shf.ui.theme.SHFTheme
